2 new commandline arguments and layout
2 new commandline arguments, for the amount of half-life before expire
of the product, and the other the half-life for the isotope. That's used
to calculate the contamination and also used for calculating the
products lifetime. The layout in the report have also been updated to
use non-scientific notation, but with 8 decimacal accuracy for the
purity of the product.

The readme file, runcommandoer has also been updated with new run
commands